update jobs in list based on websockets
